refactor: move edit message components to components directory for better organization
